Repo Reorg: move feature-specific tiling functions into their feature modules
